# reduce allocations when printing stack traces (#149)