@gregwebs 2018-09-10
@lysu do you want to pull this into [our fork](https://github.com/pingcap/errors)?