@cstockton 2019-01-27
@davecheney I'm glad to see performance improvements get in, was there any improvements I could have made to the patch that could have allowed me to maintain authorship?

I was trying to follow the commits above to see if I made any mistakes, from the diff it appears you didn't want to Grow a buffer? This has a pretty big impact on performance of the patch in the common %+v scenario:

```console
BenchmarkStackFormatting/%+v-stack-10-24          15221         21093         +38.58%
BenchmarkStackFormatting/%+v-stack-30-24          28797         43941         +52.59%
BenchmarkStackFormatting/%+v-stack-60-24          28479         42015         +47.53%

BenchmarkStackFormatting/%+v-stack-10-24          6              19             +216.67%
BenchmarkStackFormatting/%+v-stack-30-24          3              33             +1000.00%
BenchmarkStackFormatting/%+v-stack-60-24          3              33             +1000.00%
```

That said I would have been happy to remove it to maintain authorship for the work I put into fixing this and patiently waiting for it to be reviewed and merged, just something to consider for your future contributors. Thanks.